+++ README.md
@@ -146,7 +146,7 @@
  ### 3단계 - 양방향을 단방향
  > 클래스 간의 방향도 중요하다. 패키지 간의 방향도 중요하다.
  - [ ] 패키지 사이의 의존 관계는 단방향이 되도록 해야 한다.
-    - [ ] MenuGroup
+    - [x] MenuGroup
     - [ ] Menu
     - [ ] Order
     - [ ] OrderTable
