+++ src/main/java/kitchenpos/menugroup/model/MenuGroup.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.menugroup.model;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
